Rework aarc_g002_entitlement library
This libary aims to handle both G002 and G069 entitlements.
For this a new G069 class was needed.
To make the code more maintainable both entitlement classes
share a common base class now.

I took the chance to change some peculiarities of the old library,
like the naming of things. As this is basically a new library
backwards compatibility was not of the utmost importance.

I also greatly expanded on the test side of the library.